requirements.txt
Django==1.6.2
MySQL-python==1.2.5
argparse==1.2.1
djangorestframework==2.3.13
pytz==2013.9
wsgiref==0.1.2
